Add "active" place functionality.
The active place, keeps track of which place is currently under investigation by the user. A click listener exists on each place UI element (list items only, map markers not yet supported), that calls a function which changes the clicked UI element's corresponding place to the "active" place. KO logic in the html changes the CSS class of the active UI element.