Change: Switch Google Places request parameters for faster request
Previously I was using map boundaries to specificy the boundaries of the Places request, but this required waiting until the map fully loaded before a request could be sent. Instead I now use a location and radius parameter, allowing the request to be sent even before the map has loaded.